refactor: CoreData code cleanup
Resolves: none.
