+++ GitIt/Controller/Class/ExploreViewController.swift
@@ -9,11 +9,8 @@ import UIKit
 
 class ExploreViewController: UITableViewController {
     
-    var dataController: DataController!
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-
     }
 
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
