+++ GitIt/Controller/Logic/Repository/RepositoryDetailLogicController.swift
@@ -60,12 +60,12 @@ class RepositoryDetailLogicController {
     func bookmark(then handler: @escaping ViewStateHandler) {
         defer { handler(.bookmarked) }
         if !isBookmarked {
-            guard DataController.standard.insert(model) != nil else {
+            guard CoreDataManager.standard.insert(model) != nil else {
                 isBookmarked = true
                 return
             }
         } else {
-            guard DataController.standard.delete(model) != nil else {
+            guard CoreDataManager.standard.delete(model) != nil else {
                 isBookmarked = false
                 return
             }
@@ -75,7 +75,7 @@ class RepositoryDetailLogicController {
     func checkIfStarredOrBookmarked(then handler: @escaping ViewStateHandler) {
         NetworkClient.standard.checkAuthenticatedUserDidStar(fullName: model.fullName) { error in
             defer { handler(.presenting) }
-            let fetchResult = DataController.standard.exists(self.model)
+            let fetchResult = CoreDataManager.standard.exists(self.model)
             switch fetchResult {
             case .success(let exists): self.isBookmarked = exists
             case .failure(_): self.isBookmarked = false
