+++ GitIt/Controller/Logic/User/UserDetailLogicController.swift
@@ -56,12 +56,12 @@ class UserDetailLogicController {
     func bookmark(then handler: @escaping ViewStateHandler) {
         defer { handler(.bookmarked) }
         if !isBookmarked {
-            guard DataController.standard.insert(model) != nil else {
+            guard CoreDataManager.standard.insert(model) != nil else {
                 isBookmarked = true
                 return
             }
         } else {
-            guard DataController.standard.delete(model) != nil else {
+            guard CoreDataManager.standard.delete(model) != nil else {
                 isBookmarked = false
                 return
             }
@@ -71,7 +71,7 @@ class UserDetailLogicController {
     func checkIfFollowedOrBookmarked(then handler: @escaping ViewStateHandler) {
         NetworkClient.standard.checkAuthenticatedUserIsFollowing(userLogin: model.login) { error in
             defer { handler(.presenting) }
-            let fetchResult = DataController.standard.exists(self.model)
+            let fetchResult = CoreDataManager.standard.exists(self.model)
             switch fetchResult {
             case .success(let exists): self.isBookmarked = exists
             case .failure(_): self.isBookmarked = false
