+++ GitIt/Model/CommonTypes.swift
@@ -176,11 +176,11 @@ enum ContextMenuActions<Type: Model> {
     var action: UIAction {
         switch self {
             case .bookmark(let model): return UIAction(title: "Bookmark", image: UIImage(systemName: "bookmark"), identifier: nil) { action in
-                                                    _ = DataController.standard.insert(model)
+                                                    _ = CoreDataManager.standard.insert(model)
                                                 }
             
             case .unbookmark(let model): return UIAction(title: "Unbookmark", image: UIImage(systemName: "bookmark.fill"), identifier: nil) { action in
-                                                   _ = DataController.standard.delete(model)
+                                                   _ = CoreDataManager.standard.delete(model)
                                                 }
             
             case .share(let model): return UIAction(title: "Share", image: UIImage(systemName: "square.and.arrow.up"), identifier: nil) { action in
